[Crawler]: 解决出现COREDUMP的问题

+++ src/exec/monitor/mon_crwl.c
@@ -402,12 +402,12 @@ static int mon_crwl_query_worker_stat_print(crwl_cmd_t *cmd)
     fprintf(stderr, "    当前时间: %04d-%02d-%02d %02d:%02d:%02d\n"
             "    运行时长: %lu\n"
             "    平均速率: %lf\n"
-            "    当前速率: %ld\n",
+            "    当前速率: %lf\n",
             loctm.tm_year+1900, loctm.tm_mon+1, loctm.tm_mday,
             loctm.tm_hour, loctm.tm_min, loctm.tm_sec,
             stat->ctm - stat->stm,
             (double)down_webpage_total / (stat->ctm - stat->stm),
-            (down_webpage_total - last_down_webpage_total) / diff_tm);
+            (double)(down_webpage_total - last_down_webpage_total) / diff_tm);
 
     last_query_tm = stat->ctm;
     last_down_webpage_total = down_webpage_total;
